JVM-1754: [PEA] Bad graph detected in build_loop_late (#9)
After materialization, we used to replace all appearances of the old cooked
oop with the new cooked oop along JVMState. This isn't correct if the
old cooked oop is alive in the map of an ancester.

When C2 encounters a real call, it needs to aggregate debug edges from
all ancestors. If we replace the old cook oop in materization before, we
may end up with a bad graph. The late control of the new cooked oop is
higher than its early control in dominator tree.

Co-authored-by: Xin Liu <xxinliu@amazon.com>