Drop recomputation improvements
This reverts e18c0444cbe9f325d7122b97531869d2b8fedc56. Both `rank`
and `supertype` are _cached_ anyway, so the performance is not
a concern here. As for readability, while it's not really better,
it stands out from the rest of the file where everything is
recomputed on demand rather than cached and reused.

However, this adds clarifying parentheses for `rank` in `precedes`.