JVM-1617: Merge all virtual objects (#11)
* JVM-1754: [PEA] Bad graph detected in build_loop_late

After materialization, we used to replace all appearances of the old cooked
oop with the new cooked oop along JVMState. This isn't correct if the
old cooked oop is alive in the map of an ancester.

When C2 encounters a real call, it needs to aggregate debug edges from
all ancestors. If we replace the old cook oop in materization before, we
may end up with a bad graph. The late control of the new cooked oop is
higher than its early control in dominator tree.

* Enhance PEAState::print_on().

dump more information for VirtualState.

* Deep copy VirtualObject in its clone function.

This patch also changes VirtualObject::_nfields from unsigned to signed and makes
it in release build.

* Extend VirtualState to record diverging state.

if is_singleton() returns true, VirtualState is merged.
merged VirtualStates are records in _diverged.

We normalize "diverging" VirtualState. if fields are all identical, we
store 'this' rather than the real pointer of VirtualState.

* JVM-1617: [PEA] Merge all virtual objects

This patch emits PhiNode for the field which has multiple values.

* Reset virtualObject when it copys to a brand new  basic block.

* Add the unit test MergeAllVirts

* Fix release build.

* Fix linux-x86-hs-optimized build

This build is fastdebug build but also optimized, so ASSERT is not
defined. we don't define validate() if ASSERT is not defined.

* Dump more information for VirtualObject.

* Add a new test for JVM-1798.

* JVM-1798: fix VirtualObject

We insert phiNode at merging point rather than staging divergent
VirtualObjects.

* Add yet another test to cover MergeAllVirts;

---------

Co-authored-by: Xin Liu <xxinliu@amazon.com>