Add input streams that correspond to Hotspot's output streams.
An inputStream buffers a whole line at a time, no matter how long the line.
The buffer starts on stack, but uses malloc if the line size overflows.

If you use fgets or have static input buffers of fixed size, switch to this.
It is more robust and probably also more efficient.

Both newline formats of \n and \r\n are recognized.  Lines may contain nulls.
A truncated line at the end of the data is handled correctly.
There is a reliable line number counter.

The line you get back is mutable, but the stream always owns the memory.
Therefore, when you call inputStream::next, the old line is gone.
Until you call next, you can edit the line in place, and even lengthen it.

The source of data can be memory or a file or some other virtualized thing.
The input stream is seekable, if its data source is seekable.

If you need non-line-oriented input, you can still use fileStream.read.
But fileStream.readln has been removed; it is fails silently on long lines.

Line parsing over blocked input is tricky to make both efficient and correct.
This code has a gtest suite that tests all the corner cases, and tests as correct.

Possible future work:  Add format-sensitive scanning to inputStream,
so that os::print[_cr] is matched with a counterpart is::scan[_cr].
Also, if we ever use xml-log streams for input, add xml-input streams.

This line of work assumes that text files will continue to be a useful
carrier of configuration data for the HotSpot VM.  At least, it supports
the compiler oracle and CDS configuration files.  As those use cases
scale up, or for new use cases, we need a solid line reader foundation.