8303950: adding unit test for legacy window behavior
This test passes in the current JDK. The test criteria don't really reflect a rational expected behavior; they just reflect the current status quo.

This currently fails in this JDK-8303950 branch (using the new proposed AWTPaintManager), which indicates we've changed behavior.

In this case it's "window 4" that's failing. We used to get a mostly blue background, and now we're getting a background that's mostly red with a little blue. This isn't necessarily a bad thing; this failure is just documenting a change.