8303904: rolling back prev createAcceleratedImage commit
I predict once I start a code review about 8308904 that if we want to address the resolution of createAcceleratedImage: that can be a separate ticket. This change feels much more invasive to me; I wouldn't be surprised if some existing code out there is already casting the return value of createAcceleratedImage(..) as a BufferedImage.