8303904: adding failing unit test for OpenGL on Mac
I started to notice some intermittent failures in the Robot test, so now I'm having it wait a few beats on the event dispatch thread before grabbing the pixel color.

Also I'm adding a more abstract check-the-translucency-of-the-BufferedImage output. If this `testImageTransparency` passes and `testPixelColor` fails: that's a red flag that maybe the screen isn't ready when the Robot grabs the pixel.