4200096: OffScreenImageSource.removeConsumer NullPointerException
Removing a consumer (which can happen implicitly if an ImageObserve returns false) could cause a NPE in OffScreenImageSource. This NPE was caught, but it was printed to System.err.

This came to my attention when a particular set of steps flooded our log with NPEs that were actually harmless/meaningless.

The new attached test case includes two related tests: one mimics the NPE I observed, and the other is more aligned with the original wording in ticket JDK-4200096.