4200096: OffScreenImageSource.removeConsumer NullPointerException
Make sure OffScreenImageSource#addConsumer doesn't throw a NPE if the argument is null.

The rationale here is:
The preexisting implementation wouldn't throw a NPE, so we shouldn't change that now.

(Or more specifically: the preexisting implementation *would* throw a NPE, but it would also catch it and print it to System.err. The caller wouldn't need to anticipate a NPE.)