4200096: rewrite solution, add LegitimateNullPointerTest
Also I'm breaking out the unit tests (now there are 4 of them) into their own directory.

The new LegitimateNullPointerTest confirms that we DO want to see NullPointerExceptions printed to System.err IF they come from the ImageConsumer itself.

In this ticket most of our focus has been on the NPE's that stem from removing an ImageConsumer from OSIS mid-production (so the NPE was when OSIS tried to interact with its `theConsumer` field). This test tries to add other possible NPE's to our consideration.

This new test passed in the master branch before this branch. I want to preserve this existing behavior if this proposal is accepted.