4200096: rewrite solution, only print stacktrace if theConsumer != null
I don't like that we're *expecting* a NullPointerException to get thrown in the course of an allowed use case, but that's probably just a preferences on my part. More importantly: this branch passes all new unit tests. Plus it's readable and easy to code review.

This is related to the latest round of PR feedback here:
https://github.com/openjdk/jdk/pull/13408#issuecomment-1504605358