8305995: Use full dominant search for regions
This is a fix for the regression introduced by
da43cb5e463069cf4dafb262664f0d3d7c2e0eac in fix 8224957.

This regression was found while attempting to migrate an application
from JDK 1.8 to JDK 17, by running internal benchmarks, and while
investigating abnormal memory usage for about 4 times more from one of
them.

The regression appears in provided JMH benchmark, which builds a RB-tree
based map which contains 780 entries with primitive integers.

This benchmark was run with `ParallelGC` on different JVMs: `JDK
1.8.0_362`, `JDK 11.0.18`, `JDK 13.0.13`, `JDK 15.0.9`, `JDK 17.0.6`,
`JDK 19.0.2` and `JDK 20`. This allows to see that something has changed
between 13 and 15, and that the memory footprint for this code has
increased from nothing `≈ 10⁻³ ` to `7536` bytes per operation.

Proposed fix reduces the memory footprint expected value.