Calculate garbage bytes for evacuation failed regions from marked live bytes
During post_evacuate_phase1/2 we need to determine whether a given evacuation failed region
should be retained (kept for evacuation) or not. We need the liveness information for that.
However at the moment we calculate liveness/garbage during post_evac_phase1, which is too
late as we already need it during that phase.

For this reason this change collects the sum of live bytes in evacuation failed regions in
G1ConcurrentMark's _live_bytes array, using the normal mechanism to do so (with cache etc).
This value will be used in the next change.

This necessitates two main changes:
  * After copying objects, before post_evacuate_phase1 we need to flush these caches. This
    is a fairly constant time operation.
  * the _live_bytes value for regions needs to be cleared more often, just like we use the
    bitmap for evacuation failure.

Alternative options were considered (like keeping live bytes in the PSS'es, per failed
region, or adding another phase somewhere), but ultimately this seemed to be the least
invasive option.
I.e. what the marking already implements would have to be duplicated in some form or other.