6988055: Update java.net.IDN to UTS#46 and merge parts of ICU.
With this change, the static functions in java.net.IDN match the DNS usage
in the big three web browsers, Go, PHP, Python and whatever.

The ICU merge includes the classes that had to be merged in order to make
UTS#46 work, with the smallest reasonable diff:
 - ICUUncheckedIOException was replaced by UncheckedIOException
 - Source code changed to ASCII with \u escapes as needed
 - Tweaks wrt. deprecation warnings (ICU uses that for internal code)
 - No Oracle/OpenJDK license boilerplate added
 - No attempt at eliminating dead functions

The original UTS46.java contains some code to help registries. This code
has been removed in this version of UTS46.java (registries use ICU
directly anyway).

It would be possible to remove much dead code, e.g. the ULocale system
should be removable. I have not done that in order to keep down the size
of the diff between the original ICU files and the copies in JDK.