8309499: Don't swallow diagnostics with annotation processors.
Within `initProcessAnnotations` a `deferredDiagnosticHandler` is
installed when annotation processing is enabled. Normally, when
`processAnnotations` is called after `enterTrees`, the
`processAnnotations` code calls `reportDeferredDiagnostics`. However, if
an Abort gets thrown in `enterTrees` (currently only when java.lang is
missing), no one ends up calling `reportDeferredDiagnostics` and these
diagnostics end up being lost.

This commit adds a call in the finally clause to catch this situation
and adds tests based on the existing NoJavaLang tests to ensure that
these errors bubble up even when annotation processors are enabled.