JDK-8309322: [GenShen] TestAllocOutOfMemory#large failed
When generational Shenandoah is used, there may be an additional
alignment related heap size adjustment that the test should be cognizant
of. Such alignment might also happen in the non-generational case, but
in this case the specific size used in the test was affected on machines
with larger than usual os page size settings.

The alignment related adjustment would have affected all generational
collectors (except perhaps Gen Z). In the future, we might try and relax
this alignment constraint.alignment.