8308762: Metaspace leak with Instrumentation.retransform
Fix a small leak in constant pool merging during retransformation of
a class. If this class has a catch block with Throwable, the class
Throwable is pre-resolved in the constant pool, while all the other
classes are in a unresolved state. So the constant pool merging
process was considering the entry with pre-resolved class as different
compared to the destination and create a new entry. We now try to
consider it as equal specially for Methodref/Fieldref.