Add a PEAParanoid mode. (#52)
* Add a PEAParanoid mode.

In this mode, C2 aborts when it is expanding an object which has been
materialized at least once. This is not supposed to happen and will have
allocation duplication on some paths.

There are two reasons for this. 1) because we skip passive
materialization, an object PHI may render the object NSR and
hinders it from eliminating. if it's a reducible phi, try to use Cesar's
patch to mitigate it. 2) we don't support recursive materialization
like Graal PEA. I added a new testcase 'CompositeObjects' to demonstrate
it.

* fix formatting issue

---------

Co-authored-by: Xin Liu <xxinliu@amazon.com>