8311691: C2: Remove legacy code related to PostLoopMultiversioning
As discussed in JDK-8308994, we are working on re-implementation of post
loop vectorization and planning to refactor current SuperWord code. As
nobody is using or maintaining the old implementation now, to make the
refactoring work easier, we propose to remove the legacy code of the old
implementation first.

This patch removes all code realted to `PostLoopMultiversioning` inside
and outside SuperWord. After the removal, `SLP_extract()` in SuperWord
should only work on main loops. So we also removed all `is_main_loop()`
checks inside and added assertions instead.

Tested with hotspot::hotspot_all_no_apps, jdk tier1~3, langtools tier1
and 100k fuzzer tests on x86 and AArch64, no issue is found.