8311130: AArch64: Sync SVE related CPU features with VM options
As discussed in PR #14533, keeping AArch64 flag `UseSVE` and its related
CPU features in sync helps to simplify rules in IR tests. In this patch,
we mask SVE related CPU features off if specified SVE level in VM option
is lower than the hardware supported. Also, to support this change, we
move the features string construction to the end of the `initialize()`
function.

We also revert IR rule changes in PR #14533 and fix some code styles. We
tested almost full jtreg on SVE, SVE2 and non-SVE CPUs and no new issue
is found after this patch.