8309697: [TESTBUG] Remove "@requires vm.flagless" from jtreg vectorization tests
This patch removes `@require vm.flagless` annotations from HotSpot jtreg
tests in `compiler/vectorization/runner`. All jtreg cases in this folder
are invoked by test driver `VectorizationTestRunner.java` which checks
both correctness and vectorizability (IR) for each test method. We added
flagless requirement before because extra flags may mess with compiler
control in the test driver for correctness check. But `flagless` has a
side effect that it makes tests with extra flags skipped. So we propose
to get rid of it now.

To adapt the removal of `@require vm.flagless`, a few checks are added
in the test driver to skip the correctness check if extra flags make the
compiler control not work. This patch also moves previously hard-coded
flag `-XX:-OptimizeFill` in the test driver to conditions in IR checks.

Tested various of compiler control related VM flags on x86 and AArch64.