8313678 - SymbolTable can leak Symbols during cleanup
The cleanup routine of concurrentHashTable uses the lookup
function for its secondary effect of checking whether any
particular value is dead or not. It does not actually care about
the main purpose of the lookup function, i.e. checking if a particular
value is the desired value.

The lookup function for SymbolTable, however, has the side effect of
incrementing the Symbol refcount when the lookup succeeds. This is
presumably with the assumption that a successful lookup will result
in a newly held reference. concurrentHashTable's delete_in_bucket
can break this assumption by completing a successful lookup without
retaining the result (nor ever decrementing the refcount).

Thus, with a particular sequence of events (as shown in the new
test case), a newly inserted Symbol can have a refcount of 2 instead
of 1. Even if all legitimate references to this Symbol are removed,
it will remain live, thus leaking.

The fix allows the caller of .equals to specify if they intend to
use the value after the lookup, or not, allowing SymbolTable to
only increment the refcount when appropriate.