6415065: Submenu is shown on wrong screen in multiple monitor environment
What happens here is that when the position of a popup menu is
calculated, it can expand above or below, depending on the position of
the parent menu, the item being expanded and the size of the submenu and
screen resolution.

If the menu decides to expand above, the position calculation in
JMenu.getPopupMenuOrigin may yield a coordinate above the current
screen.  Later, JPopupMenu.adjustPopupLocationToFitScreen tries to fit
the entire menu into the screen.  However, it has no idea which screen
is correct, as all it has is an (x, y) location. If that location is
invalid, it may correct it by fitting it into the screen. However, if it
is valid, but located on an incorrect screen, then the whole logic goes
awry and the menu is fitted into the wrong screen.

In theory, a similar issue is possible if the menu decides to expand
below, though it's not easily reproducible as it would require a rather
unique combination of factors, such as unusually large
Menu.submenuPopupOffsetY.

The same thing could also happen to the X location, although it's less
likely because the popup menu has to be unusually wide for that.

Fix by pre-adjusting the location to fit into the correct screen in
JMenu.getPopupMenuOrigin, where the correct screen is still known.  The
resulting location may still not be final, as the menu's height needs to
be taken into account as well, but that's exactly what
JPopupMenu.adjustPopupLocationToFitScreen does.  Since the coordinate is
on the correct screen now, it fits the menu into the same screen, which
guarantees it'll be the correct one.