8315576: compiler/codecache/CodeCacheFullCountTest.java fails after JDK-8314837
CodeCacheFullCountTest disables code cache flushing, loads classes,
makes the code cache running out of space for accommodating another
nmethod, and disables compilers.
The test expects two scenarios: 1) the process continues and terminates
normally or 2) the adapters cannot be allocated in `refClass` and
VirtualMachineError has been thrown and printed.

When the test runs with option `-Xcomp` and `-XX:-TieredCompilation`,
more codes are compiled and the code cache becomes full quickly. The
adapters may not be allocated in `loadClass`, and VirtualMachineError
has been thrown but wrapped in LambdaConversionException
(`InnerClassLambdaMetafactory.buildCallSite`) and BootstrapMethodError
(`BootstrapMethodInvoker.invoke`).
VirtualMachineError may not be completely printed by `printStackTrace`
in `ThreadGroup.uncaughtException` before the process terminated. This
breaks the expected scenario 2.

The check of scenario 2 is removed from the test as it's not reliable.

The test has passed as below.
```
make test \
    TEST="test/hotspot/jtreg/compiler/codecache/CodeCacheFullCountTest.java" \
    JTREG="JAVA_OPTIONS=-Xcomp -ea -esa -XX:-TieredCompilation"
```